+++ ofdft_ml/ext_math.py
@@ -37,7 +37,6 @@ def cholesky_solver(L, b):
 def euclidean_distance(X, Y):
     # special care about complex entry
     X_ = X[:, np.newaxis, :]
-    Y_ = Y[np.newaxis, :, :]
-    D_ = X_ - Y_
+    D_ = X_ - Y
     distance = np.sum(D_*D_.conj(), axis=2, dtype=np.float64)
     return np.sqrt(distance)